Consolidate trace enrichment helpers and drop containsKey guards
- addContextInfo (repo/branch/user) now has one definition: a public
  static helper on BaseWebHandler, next to addTraceInfo. The three private
  copies in TraceWebFilter, AdminToolHandler and HandlerHelper are gone;
  IdentityHandler calls the shared one.
- Single-use private helpers are inlined: WebAppHandler.addTraceInfo and
  SlashApiHandler's misspelled addTranceInfo.
- The !trace.containsKey("app") guards in the websocket-endpoint factories
  are removed: attributes are overwrite-able by design (last writer wins,
  and the value is the same application key in every path that could have
  set it earlier), and the guard read the trace through the deprecated Map
  view. MappingHandlerWorker records the possibly-null app null-safely -
  the null-ignoring attribute writer drops it.
