+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/handler/HandlerHelper.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.Collection;
 import java.util.EnumSet;
 import java.util.Map;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 import com.google.common.io.ByteSource;
@@ -13,14 +12,11 @@ import com.google.common.primitives.Longs;
 
 import com.enonic.xp.app.ApplicationKey;
 import com.enonic.xp.branch.Branch;
-import com.enonic.xp.context.Context;
-import com.enonic.xp.context.ContextAccessor;
 import com.enonic.xp.descriptor.DescriptorKey;
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
 import com.enonic.xp.project.ProjectName;
 import com.enonic.xp.resource.Resource;
-import com.enonic.xp.security.auth.AuthenticationInfo;
 import com.enonic.xp.trace.Trace;
 import com.enonic.xp.web.HttpMethod;
 import com.enonic.xp.web.HttpStatus;
@@ -139,18 +135,6 @@ public final class HandlerHelper
             trace.attribute( "status", webResponse.getStatus().value() );
             trace.attribute( "type", webResponse.getContentType().toString() );
             trace.attribute( "size", getSize( webResponse ) );
-        }
-    }
-
-    public static void addContextInfo( final Trace trace )
-    {
-        final Context context = ContextAccessor.current();
-        trace.attribute( "repo", Objects.toString( context.getRepositoryId(), null ) );
-        trace.attribute( "branch", Objects.toString( context.getBranch(), null ) );
-        final AuthenticationInfo authInfo = context.getAuthInfo();
-        if ( authInfo != null && authInfo.getUser() != null )
-        {
-            trace.attribute( "user", authInfo.getUser().getKey().toString() );
         }
     }
